Move circuit topology check to DeviceSpecs.compile().
