+++ strawberryfields/devicespecs/__init__.py
@@ -65,15 +65,9 @@ from .gbs import GBSSpecs
 from .tensorflow import TFSpecs
 
 
-backend_specs = {
-    "base": BaseSpecs,
-    "chip0": Chip0Specs,
-    "fock": FockSpecs,
-    "gaussian": GaussianSpecs,
-    "gbs": GBSSpecs,
-    "tf": TFSpecs,
-}
-"""dict[str, DeviceSpecs]: dictionary mapping backend shortname
-to the validation dataclass."""
+devices = (BaseSpecs, Chip0Specs, FockSpecs, GaussianSpecs, GBSSpecs, TFSpecs)
+
+backend_specs = {c.short_name: c for c in devices}
+"""dict[str, DeviceSpecs]: dictionary mapping device short_name to the corresponding class."""
 
 __all__ = ["backend_specs", "DeviceSpecs"]
