+++ strawberryfields/devicespecs/tensorflow.py
@@ -18,6 +18,7 @@ from .device_specs import DeviceSpecs
 class TFSpecs(DeviceSpecs):
     """Validation data for the TF backend"""
 
+    short_name = 'tf'
     modes = None
     local = True
     remote = False
